report-automation-with-python/requirements.txt
pandas
XlsxWriter
pyodbc
PyYAML
azure-storage-file-share